+++ Recipes-API/Recipes-API/Models/RecipeIngredient.cs
@@ -8,13 +8,14 @@ namespace Recipes_API.Models;
 public partial class RecipeIngredient
 {
     [JsonIgnore]
-    public long Recipe { get; set; }
+    public int Recipe { get; set; }
 
-    public long Ingredient { get; set; }
+    public int Ingredient { get; set; }
 
-    public long Amount { get; set; }
+    public int Amount { get; set; }
 
     public virtual Ingredient IngredientNavigation { get; set; } = null!;
+
     [JsonIgnore]
     public virtual Recipe RecipeNavigation { get; set; } = null!;
 }
