+++ Recipes-API/Recipes-API/Repositories/NationalCuisineRepository.cs
@@ -19,7 +19,7 @@ public class NationalCuisineRepository
         return await dbContext.NationalCuisines.ToListAsync();
     }
 
-    public async Task<long> GetIdOrAddAsync(string nationalCuisine)
+    public async Task<int> GetIdOrAddAsync(string nationalCuisine)
     {
         var n_nationalCuisine = await dbContext.NationalCuisines.FirstOrDefaultAsync((x) => x.Name == nationalCuisine);
         if (n_nationalCuisine != null)
